feat: [BBND-1498] create Dividend port-in and remove dividend methods from Equity
